初步完成生成健康报告功能

+++ src/main/java/com/nurturing/DTO/UserPageCenterDataDTO.java
@@ -5,19 +5,20 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.text.DecimalFormat;
 
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class UserPageCenterDataDTO {
-    private int user_id;
+    private Long user_id;
     private String Username;
     private String gender;
     private int Age;
     private String phone_number;
-    private int Weight;
-    private int Height;
+    private BigDecimal Weight;
+    private BigDecimal Height;
     private String family_history;
     private String allergy_history;
     private String past_medical_history;
